+++ frontend/src/assets/GraphicalSelector/step1.jpg
