+++ frontend/src/assets/GraphicalSelector/step2.jpg
