+++ frontend/src/assets/GraphicalSelector/step3.jpg
