+++ frontend/src/assets/LandingPage/dragIcon.png
