+++ frontend/src/assets/TermPlannerHelp/additional-options.gif
