+++ frontend/src/assets/TermPlannerHelp/hide-year.gif
