+++ frontend/src/pages/LandingPage/InteractiveViewSection/index.ts
@@ -1,0 +1,3 @@
+import InteractiveViewSection from './InteractiveViewSection';
+
+export default InteractiveViewSection;
