+++ frontend/src/pages/LandingPage/KeyFeaturesSection/BlobBackground.tsx
@@ -1,0 +1,27 @@
+import React from 'react';
+import S from './styles';
+
+const BlobBackground = () => {
+  return (
+    <S.BlobContainer>
+      <svg
+        width="1000"
+        height="500"
+        viewBox="0 0 1574 634"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <path
+          d="M555.575 144.643C92.2393 105.86 139.058 465.275 305.508 586.025C563.556 773.224 795.177 342.691 1221.21 479.582C1647.25 616.473 1654.06 88 1418.8 88C1183.55 88 1007.32 182.455 555.575 144.643Z"
+          fill="#EFDBFE"
+        />
+        <path
+          d="M387.493 57.0908C-100.95 18.3438 -51.5946 377.43 123.875 498.069C395.908 685.097 640.079 254.958 1089.2 391.724C1538.33 528.49 1545.5 0.500032 1297.5 0.5C1049.5 0.499968 863.722 94.8689 387.493 57.0908Z"
+          fill="#E5C2FF"
+        />
+      </svg>
+    </S.BlobContainer>
+  );
+};
+
+export default BlobBackground;
