+++ frontend/src/pages/LandingPage/KeyFeaturesSection/index.ts
@@ -1,0 +1,3 @@
+import KeyFeaturesSection from './KeyFeaturesSection';
+
+export default KeyFeaturesSection;
